# api_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRIA v10 API",
    "description": "AI-powered school management system — 12 motor types for generating pedagogical materials, curriculum management, and scheduling. JWT authentication.",
    "version": "10.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "description": "Authenticate with username/password. Returns JWT token and user info.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register User",
        "description": "Create a new user. Admin only.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get Current User",
        "description": "Returns the currently authenticated user's info.",
        "operationId": "get_me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["auth"],
        "summary": "Update Current User",
        "description": "Update current user settings (e.g., student_book preference).",
        "operationId": "update_me_api_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/materials": {
      "get": {
        "tags": ["materials"],
        "summary": "List Materials",
        "description": "List all uploaded materials for the current user.",
        "operationId": "list_materials_api_materials_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["materials"],
        "summary": "Upload Material",
        "description": "Upload a material file (PDF, DOCX, TXT). Stores file and metadata. Max 50MB.",
        "operationId": "upload_material_api_materials_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/MaterialUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/materials/{id}": {
      "delete": {
        "tags": ["materials"],
        "summary": "Delete Material",
        "description": "Delete a material and its associated file.",
        "operationId": "delete_material_api_materials__id_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Material Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/materials/{id}/output": {
      "get": {
        "tags": ["materials"],
        "summary": "Get Material Output",
        "description": "Get full result_json from a motor result by ID.",
        "operationId": "get_material_output_api_materials__id_output_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Motor Result Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Motor result JSON",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/curriculums": {
      "get": {
        "tags": ["curriculums"],
        "summary": "List Curriculums",
        "description": "List all curriculums for the current user.",
        "operationId": "list_curriculums_api_curriculums_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["curriculums"],
        "summary": "Create Curriculum",
        "description": "Create a new curriculum from uploaded material.",
        "operationId": "create_curriculum_api_curriculums_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurriculumCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/curriculums/latest": {
      "get": {
        "tags": ["curriculums"],
        "summary": "Get Latest Curriculum",
        "description": "Get the most recently created curriculum.",
        "operationId": "get_latest_curriculum_api_curriculums_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurriculumResponse"
                }
              }
            }
          },
          "404": {
            "description": "No curriculum found"
          }
        }
      }
    },
    "/api/curriculums/{id}": {
      "delete": {
        "tags": ["curriculums"],
        "summary": "Delete Curriculum",
        "description": "Delete a curriculum by ID.",
        "operationId": "delete_curriculum_api_curriculums__id_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Curriculum Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": { "type": "string", "title": "Error message" }
        },
        "type": "object",
        "required": ["error"]
      },
      "SuccessResponse": {
        "properties": {
          "data": {
            "properties": {
              "deleted": { "type": "boolean" },
              "updated": { "type": "boolean" },
              "reset": { "type": "boolean" },
              "cleared": { "type": "boolean" },
              "success": { "type": "boolean" }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "username": { "type": "string", "minLength": 1, "title": "Username" },
          "password": { "type": "string", "minLength": 1, "title": "Password" }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "data": {
            "properties": {
              "token": { "type": "string", "title": "JWT access token" },
              "user": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "nombre": { "type": "string" },
                  "role": { "type": "string" },
                  "nivel": { "type": "string" },
                  "grado": { "type": "string" }
                },
                "required": ["id", "nombre", "role"]
              }
            },
            "required": ["token", "user"],
            "type": "object"
          }
        },
        "type": "object",
        "required": ["data"]
      },
      "RegisterRequest": {
        "properties": {
          "username": { "type": "string", "minLength": 3, "maxLength": 50, "pattern": "^[a-zA-Z0-9_]+$", "title": "Username" },
          "password": { "type": "string", "minLength": 6, "maxLength": 128, "title": "Password" },
          "nombre": { "type": "string", "minLength": 2, "maxLength": 100, "title": "Full name" },
          "nivel": { "type": "string", "enum": ["Primaria", "Secundaria"], "title": "Level" },
          "grado": { "type": "string", "title": "Grade" }
        },
        "type": "object",
        "required": ["username", "password", "nombre"]
      },
      "UpdateMeRequest": {
        "properties": {
          "student_book": { "type": "boolean", "title": "Has student book" }
        },
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "title": "User ID" },
              "nombre": { "type": "string", "title": "Full name" },
              "role": { "type": "string", "title": "Role (admin/teacher)" },
              "nivel": { "type": "string", "title": "Level" },
              "grado": { "type": "string", "title": "Grade" }
            },
            "required": ["id", "nombre", "role"]
          }
        },
        "type": "object"
      },
      "UserCreateResponse": {
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "created": { "type": "string", "format": "date-time" }
            }
          }
        },
        "type": "object"
      },
...
      "MaterialUploadRequest": {
        "properties": {
          "file": { "type": "string", "format": "binary", "title": "PDF/DOCX/TXT file (max 50MB)" },
          "tipo": { "type": "string", "title": "Material type (default: textbook)" }
        },
        "type": "object",
        "required": ["file"]
      },
      "MaterialUploadResponse": {
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "filepath": { "type": "string", "nullable": true },
              "created": { "type": "string", "format": "date-time" }
            }
          }
        },
        "type": "object"
      },
      "MaterialsListResponse": {
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "filename": { "type": "string" },
                "tipo": { "type": "string" },
                "size": { "type": "integer" },
                "filepath": { "type": "string", "nullable": true },
                "created_at": { "type": "string", "format": "date-time" }
              }
            }
          }
        },
        "type": "object"
      },
      "CurriculumCreateRequest": {
        "properties": {
          "material_id": { "type": "integer", "nullable": true },
          "unidad_real": { "type": "string" },
          "temas": { "type": "array", "items": {} },
          "contenido_temas": { "type": "array", "items": {} },
          "paginas_temas": { "type": "array", "items": {} },
          "raw_text": { "type": "string", "default": "" }
        },
        "type": "object",
        "required": ["unidad_real"]
      },
      "CurriculumCreateResponse": {
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" }
            }
          }
        },
        "type": "object"
      },
      "CurriculumResponse": {
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "material_id": { "type": "integer", "nullable": true },
              "unidad_real": { "type": "string" },
              "temas": { "type": "array", "items": {} },
              "contenido_temas": { "type": "array", "items": {} },
              "paginas_temas": { "type": "array", "items": {} },
              "raw_text": { "type": "string" },
              "created_at": { "type": "string", "format": "date-time" }
            }
          }
        },
        "type": "object"
      },
      "CurriculumsListResponse": {
        "properties": {
          "data": {
            "type": "array",
            "items": {}
          }
        },
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}